fix(ingest): string_db — add StringConfig::new(org_id) default constructor, fix min_combined_score type to u16
- Rename 3-arg constructor to `with_params(species_id, min_combined_score, org_id)`
- Add `new(org_id: Uuid) -> Self` with defaults: species_id=9606, min_combined_score=400
- Change `min_combined_score` field type from i16 to u16
- Cast combined_score to u16 in runner.rs for threshold comparison (scores are non-negative 0-1000)

+++ crates/bdp-ingest/src/pipelines/string_db/runner.rs
@@ -62,7 +62,7 @@ impl PipelineRunner for StringPipelineRunner {
                     continue;
                 }
             };
-            if row.combined_score < min_score {
+            if (row.combined_score as u16) < min_score {
                 continue;
             }
             if !should_keep(&row.protein1, &row.protein2) {
